Generate a puzzle from the provided answer key
I can't believe it was this simple!

The `GeneratePuzzleFromAnswerKey` function took me a few tries, but I successfully managed to get it to generate me a 3D Sudoku!

...Maybe one day I'll actually solve one.

I realize that I could adjust the puzzle generation function to generate puzzles of varying difficulty, but I'm satisfied for now just generating a puzzle.